fix(escalation): Little fixes so that g++ compiles on centos7
Other changes:
* id are 64 unsigned integers.
* Factoriation of service/host code.
* New attributes accessible from getters/setters.
* Code clean.

+++ src/hostgroup.cc
@@ -70,11 +70,11 @@ hostgroup::hostgroup(std::string const& name,
   _notes_url = notes_url;
 }
 
-unsigned int hostgroup::get_id() const {
+uint64_t hostgroup::get_id() const {
   return _id;
 }
 
-void hostgroup::set_id(unsigned int id)
+void hostgroup::set_id(uint64_t id)
 {
   _id = id;
 }
